pyproject.toml
[project]
name = "blackstoriesejercicio"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-generativeai",
    "anthropic",
    "ollama",
    "python-dotenv",
]
